Refactor GFN client into focused modules
